requirements.txt
certifi==2019.9.11
Django==2.2.6
django-appconf==1.0.3
django-ckeditor==5.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
Markdown==3.1.1
markdown2==2.3.8
pilkit==2.0
Pillow==6.2.0
pipenv==2018.11.26
pytz==2019.3
shortid8==1.0
six==1.12.0
sqlparse==0.3.0
ueditor==1.4
virtualenv==16.7.5
virtualenv-clone==0.5.3
